# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ginlol",
        "title": "Ginlol",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/": {
            "get": {
                "description": "帳號密碼",
                "tags": [
                    "Hello"
                ],
                "summary": "初始",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/demo/v1/hello/{user}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hello"
                ],
                "summary": "說HALO",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名字",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "login"
                ],
                "summary": "呈現登入頁面"
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json."
                ],
                "tags": [
                    "login"
                ],
                "summary": "\"帳號密碼輸入,如果沒有就新增\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User struct",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User struct",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User struct",
                        "name": "password-again",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": \"You are logged in!\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\": \"unauthorized\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.IndexData": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}